Add smart component for timeline controller.
